+++ src/modules/flow/initFlowSystem.ts
@@ -16,6 +16,7 @@ import { registerBuiltInNodes } from "./types/RegisterBuiltInNodes.ts";
 import { register } from "./types/Example/customNode.js";
 import { register as registerMathNode } from "./types/Nodes/mathExpressionNode";
 import { register as registerTextNode } from "./types/Nodes/textTemplateNode";
+import { register as registerGroqChatNode } from "./types/Nodes/GroqChatNode.ts";
 
 import { nodeRegistry } from "./types/NodeRegistry.ts";
 
@@ -30,6 +31,7 @@ export async function initFlowSystem() {
   registerMathNode(nodeRegistry);
 
   registerTextNode(nodeRegistry);
+  registerGroqChatNode(nodeRegistry);
 }
 
 export async function loadModule(name: string) {
