+++ src/modules/flow/types/NodeTypes.ts
@@ -107,12 +107,6 @@ export interface NumberNode extends BaseNode {
   process: (context: NodeExecutionContext) => Promise<NodeValue | undefined>;
 }
 
-export interface ChatNode extends BaseNode {
-  nodeType: "Chat";
-  nodeValue?: NodeValue;
-  process: (context: NodeExecutionContext) => Promise<NodeValue | undefined>;
-}
-
 export interface BooleanNode extends BaseNode {
   nodeType: "Boolean";
   nodeValue?: NodeValue;
@@ -148,7 +142,6 @@ export type NodeType =
   | BaseNode
   | TextNode
   | NumberNode
-  | ChatNode
   | BooleanNode
   | ImageNode
   | GenericNode
@@ -197,201 +190,6 @@ export const createNumberNode = (
       UIConfigurable: true,
       description: "Constant number to output",
       isNodeBodyContent: true,
-    },
-  ],
-  getConfigParameters: function () {
-    return this.configParameters || [];
-  },
-  getConfigParameter(parameterName) {
-    const parameter = (this.configParameters ?? []).find(
-      (param) => param.parameterName === parameterName
-    );
-    return parameter;
-  },
-  setConfigParameter(parameterName, value) {
-    const parameter = (this.configParameters ?? []).find(
-      (param) => param.parameterName === parameterName
-    );
-    if (parameter) {
-      parameter.paramValue = value;
-    }
-  },
-});
-
-export const createGroqChatNode = (
-  id: number,
-  type: { x: number; y: number }
-): ChatNode => ({
-  id,
-  title: "Chat",
-  nodeValue: "Model:",
-  nodeType: "Chat",
-  sockets: [
-    {
-      id: id * 100 + 1,
-      title: "Prompt",
-      type: "input",
-      nodeId: id,
-      dataType: "string",
-    },
-    {
-      id: id * 100 + 2,
-      title: "System Prompt",
-      type: "input",
-      nodeId: id,
-      dataType: "string",
-    },
-    {
-      id: id * 100 + 3,
-      title: "Response",
-      type: "output",
-      nodeId: id,
-      dataType: "string",
-    },
-    {
-      id: id * 100 + 4,
-      title: "Tokens",
-      type: "output",
-      nodeId: id,
-      dataType: "number",
-    },
-  ],
-  x: type.x,
-  y: type.y,
-  width: 300, // Wider to accommodate multiple inputs
-  height: 240, // Taller to fit multi-line text
-  selected: false,
-  processing: false,
-  process: async ({ node, getInputValue }) => {
-    const n = node as ChatNode;
-    // If we have a cached result for this node, use it
-    const promptValue = await getInputValue(n.id * 100 + 1);
-    const systemPrompt = await getInputValue(n.id * 100 + 2);
-
-    const prompt = String(promptValue || "");
-
-    // Extract model name from node value
-    const modelMatch = String(n.nodeValue || "");
-    const model = modelMatch ? modelMatch : "llama-3.1-8b-instant"; // Default fallback
-
-    // Use system prompt from input, but don't use the node.value (which now contains the model)
-    const system = String(systemPrompt || "");
-
-    try {
-      // Get API key from .env using Vite's environment variable format
-      const GROQ_API_KEY = import.meta.env.VITE_GROQ_API_KEY;
-
-      if (!GROQ_API_KEY) {
-        throw new Error(
-          "GROQ API key not found. Please check your .env file and ensure it has VITE_GROQ_API_KEY defined."
-        );
-      }
-
-      console.log(`Using model: ${model}`);
-      console.log(
-        `Executing Chat node ${n.id} with prompt: "${prompt.substring(
-          0,
-          50
-        )}..."`
-      );
-      const messages = system
-        ? [
-            { role: "user", content: prompt },
-            { role: "system", content: system },
-          ]
-        : [{ role: "user", content: prompt }];
-      const res = await fetch(
-        "https://api.groq.com/openai/v1/chat/completions",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${GROQ_API_KEY}`,
-          },
-          body: JSON.stringify({
-            model: model,
-            messages: messages,
-            max_tokens: 1000,
-            temperature: 0.7,
-            top_p: 1,
-            frequency_penalty: 0,
-            presence_penalty: 0,
-          }),
-        }
-      );
-
-      if (!res.ok) {
-        throw new Error(`Chat API returned status ${res.status}`);
-      }
-
-      const json = await res.json();
-      console.log(
-        `Chat node ${n.id} received response:`,
-        json.choices[0].message.content.substring(0, 50) + "..."
-      );
-
-      // Return an object with both values to support multiple outputs
-      return {
-        // Socket id 3 is for Response content
-        [n.id * 100 + 3]: json.choices[0].message.content,
-        // Socket id 4 is for Token count
-        [n.id * 100 + 4]: json.usage?.total_tokens || 0,
-      };
-    } catch (error) {
-      console.error("Error in Chat node:", error);
-      // Return error in the response output
-      return {
-        [n.id * 100 + 3]: `Error: ${
-          error instanceof Error ? error.message : String(error)
-        }`,
-        [n.id * 100 + 4]: 0,
-      };
-    }
-  },
-  configParameters: [
-    {
-      parameterName: "Model",
-      parameterType: "string",
-      defaultValue: "llama-3.1-8b-instant",
-      valueSource: "UserInput",
-      UIConfigurable: true,
-      description: "Initial image URL or base64 string to output",
-      isNodeBodyContent: true,
-      sourceList: [
-        {
-          key: "llama-3.1-8b-instant",
-          label: "Llama 3.1 8b Instant",
-        },
-        {
-          key: "llama-3.1-8b-instruct",
-          label: "Llama 3.1 8b Instruct",
-        },
-        {
-          key: "llama-3.1-8b-chat",
-          label: "Llama 3.1 8b Chat",
-        },
-        {
-          key: "llama-3.1-13b-instant",
-          label: "Llama 3.1 13b Instant",
-        },
-        {
-          key: "llama-3.1-13b-instruct",
-          label: "Llama 3.1 13b Instruct",
-        },
-        {
-          key: "llama-3.1-13b-chat",
-          label: "Llama 3.1 13b Chat",
-        },
-      ],
-    },
-    {
-      parameterName: "API Key",
-      parameterType: "string",
-      defaultValue: "",
-      valueSource: "UserInput",
-      UIConfigurable: true,
-      description: "Initial image URL or base64 string to output",
-      isNodeBodyContent: false,
     },
   ],
   getConfigParameters: function () {
