+++ src/modules/flow/types/RegisterBuiltInNodes.ts
@@ -14,7 +14,6 @@
 import { nodeRegistry } from "./NodeRegistry";
 import {
   createNumberNode,
-  createGroqChatNode,
   createBooleanNode,
   createImageNode,
   createAddNode,
@@ -23,7 +22,6 @@ import {
 
 export function registerBuiltInNodes(): void {
   nodeRegistry.registerNodeType("Number", createNumberNode);
-  nodeRegistry.registerNodeType("Chat", createGroqChatNode);
   nodeRegistry.registerNodeType("Boolean", createBooleanNode);
   nodeRegistry.registerNodeType("Image", createImageNode);
   nodeRegistry.registerNodeType("Add", createAddNode);
